requirements.txt
argon2-cffi==20.1.0
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
Click==7.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.1
jsonschema==3.2.0
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.18.1
packaging==20.4
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.1
pdftotext==2.1.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.22.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
